FIX: Issue with meteor publish with update option
It was missing an await in the update readme function.

Adding tests for this is an issue because of this snippet https://github.com/meteor/meteor/blob/3747d4cb875f1282121ae7df473e0b00c3b3ce7c/tools/cli/commands-packages.js#L175-L184 and other parts that require or do some kind of check with the network

Maybe for the future think a way of mocking or being able to do tests in functions and not system tests.

Increase of types could help too.